fixed: compile warnings
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@29921 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/utils/WebServer.cpp
@@ -272,7 +272,7 @@ int CWebServer::ContentReaderCallback(void *cls, size_t pos, char *buf, int max)
 #endif
 {
   CFile *file = (CFile *)cls;
-  if(pos != file->GetPosition())
+  if((unsigned int)pos != file->GetPosition())
     file->Seek(pos);
   unsigned res = file->Read(buf, max);
   if(res == 0)
